requirements.txt
# FastAPI и веб-сервер
fastapi==0.121.0
uvicorn[standard]==0.38.0

# База данных
sqlalchemy==2.0.44
asyncpg==0.30.0

# Kafka
aiokafka==0.12.0

# Pydantic и валидация
pydantic==2.12.4

# Переменные окружения
python-dotenv==1.2.1